+++ soxs/tests/test_bkgnd.py
@@ -1,10 +1,11 @@
 from soxs.instrument import make_background, \
     instrument_simulator, make_background_file, simulate_spectrum
-from soxs.background.foreground import frgnd_spec
+from soxs.background.foreground import make_frgnd_spectrum
 from soxs.background.spectra import ConvolvedBackgroundSpectrum
 from soxs.spectra import Spectrum
 from soxs.response import AuxiliaryResponseFile, RedistributionMatrixFile
-from soxs.utils import soxs_files_path
+from soxs.utils import soxs_files_path, set_soxs_config
+from soxs.tests.utils import spectrum_answer_testing
 from numpy.random import RandomState
 from numpy.testing import assert_allclose
 import astropy.io.fits as pyfits
@@ -29,7 +30,7 @@ def test_uniform_bkgnd_scale():
     fov = (event_params["fov"]*60.0)**2
     S = ncts/t_exp/fov
     dS = np.sqrt(ncts)/t_exp/fov
-    foreground = ConvolvedBackgroundSpectrum.convolve(frgnd_spec, hdxi_arf)
+    foreground = ConvolvedBackgroundSpectrum.convolve(make_frgnd_spectrum.spec, hdxi_arf)
     f_sum = foreground.get_flux_in_band(0.7, 2.0)[0]
     i_sum = acisi_particle_bkgnd.get_flux_in_band(0.7, 2.0)[0]*(u.cm/u.arcmin)**2
     b_sum = (f_sum+i_sum).to("ph/(arcsec**2*s)").value
@@ -57,7 +58,7 @@ def test_simulate_bkgnd_spectrum():
         ncts = f["SPECTRUM"].data["COUNTS"][ch_min:ch_max].sum()
     S = ncts/exp_time/fov
     dS = np.sqrt(ncts)/exp_time/fov
-    foreground = ConvolvedBackgroundSpectrum.convolve(frgnd_spec, hdxi_arf)
+    foreground = ConvolvedBackgroundSpectrum.convolve(make_frgnd_spectrum.spec, hdxi_arf)
     f_sum = foreground.get_flux_in_band(0.7, 2.0)[0]
     i_sum = acisi_particle_bkgnd.get_flux_in_band(0.7, 2.0)[0]*(u.cm/u.arcmin)**2
     b_sum = (f_sum+i_sum).to_value("ph/(arcsec**2*s)")
@@ -155,6 +156,18 @@ def test_ptsrc():
     assert_allclose(events["energy"].sum(), events2["energy"].sum(), rtol=1.0e-3)
     os.chdir(curdir)
     shutil.rmtree(tmpdir)
+
+
+def test_change_bkgnd(answer_store, answer_dir):
+    from soxs.background.foreground import make_frgnd_spectrum
+    set_soxs_config("frgnd_spec_model", "default")
+    spectrum_answer_testing(make_frgnd_spectrum.spec, 
+                            f"default_frgnd_spectrum.h5", answer_store,
+                            answer_dir)
+    set_soxs_config("frgnd_spec_model", "halosat")
+    spectrum_answer_testing(make_frgnd_spectrum.spec,
+                            f"lem_frgnd_spectrum.h5", answer_store,
+                            answer_dir)
 
 
 if __name__ == "__main__":